packages/core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-dotenv>=1.1.1"]
